update for map support
